requirements.txt
uvicorn
bareasgi
bareasgi_static
pyquery
google-api-python-client
dialogflow
PyYAML
aiofiles
redis
rom
pywsd
numpy
